move pb2 to own folder, regen script writes to new folder, update imports
