+++ bin/regen-protobufs.sh
@@ -4,7 +4,7 @@
 #gsed -i 's/import "\//import ".\//g' ./protobufs/meshtastic/*
 #gsed -i 's/package meshtastic;//g' ./protobufs/meshtastic/*
 
-./nanopb-0.4.7/generator-bin/protoc -I=protobufs --python_out ./ ./protobufs/meshtastic/*.proto
+./nanopb-0.4.7/generator-bin/protoc -I=protobufs --python_out ./pbDefinitions ./protobufs/meshtastic/*.proto
 
 # workaround for import bug in protoc https://github.com/protocolbuffers/protobuf/issues/1491#issuecomment-690618628
