+++ meshtastic/__init__.py
@@ -78,7 +78,7 @@ from google.protobuf.json_format import MessageToJson
 from pubsub import pub
 from tabulate import tabulate
 
-from meshtastic import (
+from meshtastic.pbDefinitions import (
     admin_pb2,
     apponly_pb2,
     channel_pb2,
@@ -86,10 +86,10 @@ from meshtastic import (
     mesh_pb2,
     portnums_pb2,
     remote_hardware_pb2,
-    telemetry_pb2,
     util,
 )
 from meshtastic.node import Node
+from meshtastic.pbDefinitions import telemetry_pb2
 from meshtastic.util import DeferredExecution, Timeout, catchAndIgnore, fixme, stripnl
 
 # Note: To follow PEP224, comments should be after the module variable.
