+++ meshtastic/node.py
@@ -7,7 +7,8 @@ import time
 
 from google.protobuf.json_format import MessageToJson
 
-from meshtastic import admin_pb2, apponly_pb2, channel_pb2, localonly_pb2, portnums_pb2
+from meshtastic.pbDefinitions import admin_pb2, apponly_pb2, channel_pb2, portnums_pb2
+from meshtastic.pbDefinitions import localonly_pb2
 from meshtastic.util import (
     Timeout,
     camel_to_snake,
